REBUILT WEB PACK - HOT RELOAD
Web pack now formatted in same style as the Stripe Demo - which enabled
“hot reloading,” although we have to refresh to see changes pop up.

Messed with Google Map and added a few pins, but we may need to
implement a different react module to be very efficient and placing
these. This module also does not make it easy to size things.

THINGS TO NOTE:
- Truck Entry still works, but must be accessed at
/html/truckEntry.html (web pack made me do it!)